# Make progress bars constant width