Before:

```
     Running `target\debug\probe-rs.exe run C:\_Hobby\rpi-smoke-tester-setup\embedded-test\esp-hal\output\esp32s2`
      Erasing ✔ [00:00:01] [########################################] 192.00 KiB/192.00 KiB @ 189.88 KiB/s (eta 0s )
  Programming ✔ [00:00:02] [############################################] 27.95 KiB/27.95 KiB @ 9.52 KiB/s (eta 0s )
    Finished in 2.9602304s
```

After:

```
     Running `target\debug\probe-rs.exe run C:\_Hobby\rpi-smoke-tester-setup\embedded-test\esp-hal\output\esp32s2`
      Erasing ✔ [00:00:01] [###################################] 192.00 KiB/192.00 KiB @ 189.35 KiB/s (ETA 00:00:00)
  Programming ✔ [00:00:02] [###################################]  27.95 KiB/ 27.95 KiB @   9.53 KiB/s (ETA 00:00:00)
     Finished in 2.96s
```

- Changed the display format of ETA, and Finished
- Aligned Finished with the other labels
- Widened the current/total bytes and throughput fields

Closes #2736