@bugadani 2024-09-28
Current iteration: fixed (<80) width. I've opted to display ETA while a progress bar is running, and replace it with the elapsed time once done - the jump may look weird if a stage is quick, but there is some sense in the choice I hope.

```
      Erasing ✔ 100% (00:00:01) [####################] 192.00 KiB @ 175.37 KiB/s
  Programming ⠓  33% (00:00:04) [#######-------------]  13.64 KiB @   6.50 KiB/s

      Erasing ✔ 100% (00:00:01) [####################] 192.00 KiB @ 174.74 KiB/s
  Programming ✔ 100% (00:00:03) [####################]  40.92 KiB @  11.11 KiB/s                                                                                                                                                                            
     Finished in 3.71s
```

Fun fact: indicatif allows us to have multi-line progress bars. Unfortunately what I tried looked pretty bad 🫠